@@ -1,10 +1,10 @@
 name: limiting_direction_csx
 description: Limit the rotatable orientation of the screen, also known as the orientation Settings that the interface can support. It should be noted that the orientation set here needs to be included in the supported orientation of the project.
-version: 0.0.4+3
+version: 0.1.0
 homepage: https://github.com/KirstenDunst/FlutterPlugins/tree/main/limiting_direction_csx
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.20.0"
 
 dependencies:
